requirements.txt
feedparser
beautifulsoup4
requests
python-dotenv
openai
